chore: avoid publishing tests by moving them out of src
Bring the number of published files back to the previous level.
I considered using an .npmignore file but I think this solution
requires less npm knowledge to understand.

I've also renamed `__tests__` to `test` so the layout matches the
one in the postcss and webpack repositories.
